backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pdfplumber==0.11.6
groq==0.25.0
python-dotenv==1.1.0
